requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
numpy==1.21.5
pandas==1.3.5
scikit-learn==1.0.2
joblib==1.1.0
matplotlib==3.5.1
gunicorn==20.1.0
pymongo==4.3.3
pytest==7.0.1
python-dotenv==0.19.2
flask-wtf==1.0.0
pyyaml==6.0